+++ application/models/countries_model.php
@@ -32,8 +32,6 @@ class Countries_model extends CI_Model {
 
 			if (!empty($filter['sort_by']) AND !empty($filter['order_by'])) {
 				$this->db->order_by($filter['sort_by'], $filter['order_by']);
-			} else {
-				$this->db->order_by('country_name', 'ASC');
 			}
 			
 			$query = $this->db->get();
@@ -72,39 +70,9 @@ class Countries_model extends CI_Model {
 		}
 	}
 
-	public function addCountry($add = array()) {
-
-		if (!empty($add['country_name'])) {
-			$this->db->set('country_name', $add['country_name']);
-		}
-				
-		if (!empty($add['iso_code_2'])) {
-			$this->db->set('iso_code_2', $add['iso_code_2']);
-		}
-		
-		if (!empty($add['iso_code_3'])) {
-			$this->db->set('iso_code_3', $add['iso_code_3']);
-		}
-		
-		if (!empty($add['format'])) {
-			$this->db->set('format', $add['format']);
-		}
-		
-		if ($add['status'] === '1') {
-			$this->db->set('status', $add['status']);
-		} else {
-			$this->db->set('status', '0');
-		}
-
-		$this->db->insert('countries');
-		
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
-		}
-	}
-
 	public function updateCountry($update = array()) {
-
+		$query = FALSE;
+		
 		if (!empty($update['country_name'])) {
 			$this->db->set('country_name', $update['country_name']);
 		}
@@ -129,21 +97,54 @@ class Countries_model extends CI_Model {
 
 		if (!empty($update['country_id'])) {
 			$this->db->where('country_id', $update['country_id']);
-			$this->db->update('countries'); 
+			$query = $this->db->update('countries'); 
 		}
 		
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
+		return $query;
+	}
+
+	public function addCountry($add = array()) {
+		$query = FALSE;
+
+		if (!empty($add['country_name'])) {
+			$this->db->set('country_name', $add['country_name']);
 		}
+				
+		if (!empty($add['iso_code_2'])) {
+			$this->db->set('iso_code_2', $add['iso_code_2']);
+		}
+		
+		if (!empty($add['iso_code_3'])) {
+			$this->db->set('iso_code_3', $add['iso_code_3']);
+		}
+		
+		if (!empty($add['format'])) {
+			$this->db->set('format', $add['format']);
+		}
+		
+		if ($add['status'] === '1') {
+			$this->db->set('status', $add['status']);
+		} else {
+			$this->db->set('status', '0');
+		}
+
+		if (!empty($add)) {
+			if ($this->db->insert('countries')) {
+				$query = $this->db->insert_id();
+			}
+		}
+		
+		return $query;
 	}
 
 	public function deleteCountry($country_id) {
-		$this->db->where('country_id', $country_id);
-		
-		$this->db->delete('countries');
+		if (is_numeric($country_id)) {
+			$this->db->where('country_id', $country_id);
+			$this->db->delete('countries');
 
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
+			if ($this->db->affected_rows() > 0) {
+				return TRUE;
+			}
 		}
 	}
 }
