+++ application/models/menus_model.php
@@ -50,8 +50,6 @@ class Menus_model extends CI_Model {
 			
 			if (!empty($filter['sort_by']) AND !empty($filter['order_by'])) {
 				$this->db->order_by($filter['sort_by'], $filter['order_by']);
-			} else {
-				$this->db->order_by('menu_id', 'DESC');
 			}
 		
 			if (!empty($filter['filter_search'])) {
@@ -121,7 +119,7 @@ class Menus_model extends CI_Model {
 
 	public function getAdminMenu($menu_id) {					
 		//$this->db->select('menus.menu_id, *');
-		$this->db->select('menus.menu_id, menu_name, menu_description, menu_price, menu_photo, menu_category_id, stock_qty, minimum_qty, subtract_stock, menu_status, category_id, category_name, category_description, category_special, special_id, start_date, end_date, special_price, special_status');
+		$this->db->select('menus.menu_id, menu_name, menu_description, menu_price, menu_photo, menu_category_id, stock_qty, minimum_qty, subtract_stock, menu_status, category_id, category_name, category_description, special_id, start_date, end_date, special_price, special_status');
 		$this->db->from('menus');
 		$this->db->join('categories', 'categories.category_id = menus.menu_category_id', 'left');
 		$this->db->join('menus_specials', 'menus_specials.menu_id = menus.menu_id', 'left');
@@ -144,8 +142,6 @@ class Menus_model extends CI_Model {
 		
 			if (!empty($filter['sort_by']) AND !empty($filter['order_by'])) {
 				$this->db->order_by($filter['sort_by'], $filter['order_by']);
-			} else {
-				$this->db->order_by('option_id', 'DESC');
 			}
 		
 			if (!empty($filter['filter_search'])) {
@@ -242,8 +238,6 @@ class Menus_model extends CI_Model {
 		
 			if (!empty($filter['sort_by']) AND !empty($filter['order_by'])) {
 				$this->db->order_by($filter['sort_by'], $filter['order_by']);
-			} else {
-				$this->db->order_by('category_id', 'DESC');
 			}
 		
 			if (!empty($filter['filter_search'])) {
@@ -331,7 +325,6 @@ class Menus_model extends CI_Model {
 	}
 
 	public function updateMenu($update = array()) {
-
 		$query = FALSE;
 		
 		if (!empty($update['menu_name'])) {
@@ -382,59 +375,64 @@ class Menus_model extends CI_Model {
 		
 		if (!empty($update['menu_id'])) {
 			$this->db->where('menu_id', $update['menu_id']);
-			$this->db->update('menus');
-		}
-		
-		if ($this->db->affected_rows() > 0) {
-			$query = TRUE;
-		}
 
-		$this->db->where('menu_id', $update['menu_id']);
-		$this->db->delete('menus_to_options');
+			if ($query = $this->db->update('menus')) {
+				$this->db->where('menu_id', $update['menu_id']);
+				$this->db->delete('menus_to_options');
 
-		if (!empty($update['menu_options'])) {
-			foreach ($update['menu_options'] as $key => $value) {
-				$this->db->set('menu_id', $update['menu_id']);
-				$this->db->set('option_id', $value);
-				$this->db->insert('menus_to_options');
+				if (!empty($update['menu_options'])) {
+					foreach ($update['menu_options'] as $key => $value) {
+						$this->db->set('menu_id', $update['menu_id']);
+						$this->db->set('option_id', $value);
+						$query = $this->db->insert('menus_to_options');
+					}
+				}
+
+				if (!empty($update['start_date'])) {
+					$this->db->set('start_date', mdate('%Y-%m-%d', strtotime($update['start_date'])));
+				}
+
+				if (!empty($update['end_date'])) {
+					$this->db->set('end_date', mdate('%Y-%m-%d', strtotime($update['end_date'])));
+				}
+
+				if (!empty($update['special_price'])) {
+					$this->db->set('special_price', $update['special_price']);
+				}
+
+				if ($update['special_status'] === '1') {
+					$this->db->set('special_status', '1');
+				} else {
+					$this->db->set('special_status', '0');
+				}
+			
+				if (!empty($update['special_id'])) {
+					$this->db->where('special_id', $update['special_id']);
+					$this->db->where('menu_id', $update['menu_id']);
+					$query = $this->db->update('menus_specials');
+				} else {
+					$this->db->set('menu_id', $update['menu_id']);
+					$query = $this->db->insert('menus_specials');
+				}
 			}
-		}
-
-		if ($this->db->affected_rows() > 0) {
-			$query = TRUE;
-		}
-
-		//$this->db->where('menu_id', $update['menu_id']);
-		//$this->db->delete('menus_specials');
-
-		if ($update['special_status'] === '1') {
-			if (!empty($update['start_date']) AND !empty($update['end_date']) AND !empty($update['special_price'])) {
-				$this->db->set('start_date', mdate('%Y-%m-%d', strtotime($update['start_date'])));
-				$this->db->set('end_date', mdate('%Y-%m-%d', strtotime($update['end_date'])));
-				$this->db->set('special_price', $update['special_price']);
-				$this->db->set('special_status', $update['special_status']);
-			}
-		} else {
-			$this->db->set('special_status', '0');
-		}
-				
-		if (!empty($update['special_id'])) {
-			$this->db->where('special_id', $update['special_id']);
-			$this->db->where('menu_id', $update['menu_id']);
-			$this->db->update('menus_specials');
-		} else {
-			$this->db->set('menu_id', $update['menu_id']);
-			$this->db->insert('menus_specials');
-		}
-
-		if ($this->db->affected_rows() > 0) {
-			$query = TRUE;
 		}
 		
 		return $query;
 	}
 
+	public function updateCategory($category_id, $category_name, $category_description) {
+		$this->db->set('category_name', $category_name);
+		$this->db->set('category_description', $category_description);
+		
+		$this->db->where('category_id', $category_id);
+		$query = $this->db->update('categories'); 
+
+		return $query;
+	}
+
 	public function updateMenuOption($update = array()) {
+		$query = FALSE;
+
 		if (!empty($update['option_name'])) {
 			$this->db->set('option_name', $update['option_name']);
 		}
@@ -445,27 +443,15 @@ class Menus_model extends CI_Model {
 
 		if (!empty($update['option_id'])) {
 			$this->db->where('option_id', $update['option_id']);
-			$this->db->update('menu_options'); 
+			$query = $this->db->update('menu_options'); 
 		}
 				
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
-		}
+		return $query;
 	}
 	
-	public function updateCategory($category_id, $category_name, $category_description) {
-		$this->db->set('category_name', $category_name);
-		$this->db->set('category_description', $category_description);
-		
-		$this->db->where('category_id', $category_id);
-		$this->db->update('categories'); 
-
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
-		}
-	}
-
 	public function addMenu($add = array()) {
+		$query = FALSE;
+
 		if (!empty($add['menu_name'])) {
 			$this->db->set('menu_name', $add['menu_name']);
 		}
@@ -512,41 +498,56 @@ class Menus_model extends CI_Model {
 			$this->db->set('menu_status', '0');
 		}
 		
-		$this->db->insert('menus');
-	
-		if ($this->db->affected_rows() > 0 AND $this->db->insert_id()) {
-			$menu_id = $this->db->insert_id();
+		if (!empty($add)) {
+			if ($this->db->insert('menus')) {
+				$menu_id = $this->db->insert_id();
 
-			$this->db->where('menu_id', $menu_id);
-			$this->db->delete('menus_to_options');
+				$this->db->where('menu_id', $menu_id);
+				$this->db->delete('menus_to_options');
 
-			if (!empty($add['menu_options'])) {
-				foreach ($add['menu_options'] as $key => $value) {
-					$this->db->set('menu_id', $menu_id);
-					$this->db->set('option_id', $value);
-					$this->db->insert('menus_to_options');
+				if (!empty($add['menu_options'])) {
+					foreach ($add['menu_options'] as $key => $value) {
+						$this->db->set('menu_id', $menu_id);
+						$this->db->set('option_id', $value);
+						$query = $this->db->insert('menus_to_options');
+					}
 				}
-			}
 
-			$this->db->where('menu_id', $menu_id);
-			$this->db->delete('menus_specials');
+				$this->db->where('menu_id', $menu_id);
+				$this->db->delete('menus_specials');
 
-			if ($add['special_status'] === '1') {
-				if (!empty($update['start_date']) AND !empty($add['end_date']) AND !empty($add['special_price'])) {
-					$this->db->set('menu_id', $menu_id);
-					$this->db->set('start_date', mdate('%Y-%m-%d', strtotime($add['start_date'])));
-					$this->db->set('end_date', mdate('%Y-%m-%d', strtotime($add['end_date'])));
-					$this->db->set('special_price', $add['special_price']);
-					$this->db->set('special_status', $add['special_status']);
-					$this->db->insert('menus_specials');
+				if ($add['special_status'] === '1') {
+					if (!empty($update['start_date']) AND !empty($add['end_date']) AND !empty($add['special_price'])) {
+						$this->db->set('menu_id', $menu_id);
+						$this->db->set('start_date', mdate('%Y-%m-%d', strtotime($add['start_date'])));
+						$this->db->set('end_date', mdate('%Y-%m-%d', strtotime($add['end_date'])));
+						$this->db->set('special_price', $add['special_price']);
+						$this->db->set('special_status', $add['special_status']);
+						$query = $this->db->insert('menus_specials');
+					}
 				}
+				
+				$query = $menu_id;
 			}
+			
+			return $query;
+		}
+	}
 
-			return TRUE;
+	public function addCategory($category_name, $category_description) {
+		if ($category_name OR $category_description) {
+			$this->db->set('category_name', $category_name);
+			$this->db->set('category_description', $category_description);
+		
+			if ($this->db->insert('categories')) {
+				return $this->db->insert_id();
+			}
 		}
 	}
 
 	public function addMenuOption($add = array()) {
+		$query = FALSE;
+
 		if (!empty($add['option_name'])) {
 			$this->db->set('option_name', $add['option_name']);
 		}
@@ -555,59 +556,51 @@ class Menus_model extends CI_Model {
 			$this->db->set('option_price', $add['option_price']);
 		}
 		
-		$this->db->insert('menu_options');
-
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
+		if (!empty($add)) {
+			if ($this->db->insert('menu_options')) {
+				$query = $this->db->insert_id();
+			}
 		}
-	}
 
-	public function addCategory($category_name, $category_description) {
-		
-		$this->db->set('category_name', $category_name);
-		$this->db->set('category_description', $category_description);
-		
-		$this->db->insert('categories');
-
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
-		}
+		return $query;
 	}
 
 	public function deleteMenu($menu_id) {
+		if (is_numeric($menu_id)) {
+			$this->db->where('menu_id', $menu_id);
+			$this->db->delete('menus');
 
-		$this->db->where('menu_id', $menu_id);
-		$this->db->delete('menus');
+			$this->db->where('menu_id', $menu_id);
+			$this->db->delete('menus_specials');
 
-		$this->db->where('menu_id', $menu_id);
-		$this->db->delete('menus_specials');
+			$this->db->where('menu_id', $menu_id);
+			$this->db->delete('menus_to_options');
 
-		$this->db->where('menu_id', $menu_id);
-		$this->db->delete('menus_to_options');
-
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
+			if ($this->db->affected_rows() > 0) {
+				return TRUE;
+			}
 		}
 	}
 
 	public function deleteMenuOption($option_id) {
+		if (is_numeric($option_id)) {
+			$this->db->where('option_id', $option_id);
+			$this->db->delete('menu_options');
 
-		$this->db->where('option_id', $option_id);
-
-		$this->db->delete('menu_options');
-
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
+			if ($this->db->affected_rows() > 0) {
+				return TRUE;
+			}
 		}
 	}
 
 	public function deleteCategory($category_id) {
-		$this->db->where('category_id', $category_id);
+		if (is_numeric($category_id)) {
+			$this->db->where('category_id', $category_id);
+			$this->db->delete('categories');
 
-		$this->db->delete('categories');
-
-		if ($this->db->affected_rows() > 0) {
-			return TRUE;
+			if ($this->db->affected_rows() > 0) {
+				return TRUE;
+			}
 		}
 	}
 }
